# documentation/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
...
paths:
  /v1/resetpassword:
    put:
      tags:
        - Patient
      operationId: resetPassword
      parameters:
        - name: cpr
          in: query
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
...
  /v1/patient:
    get:
      tags:
        - Patient
      operationId: getPatient
      parameters:
        - name: cpr
          in: query
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PatientDto'
    post:
      tags:
        - Patient
      operationId: createPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
...
  /v1/patients:
    get:
      tags:
        - Patient
      operationId: getPatients
      parameters:
        - name: includeActive
          in: query
          required: true
          schema:
            type: boolean
        - name: includeCompleted
          in: query
          required: true
          schema:
            type: boolean
        - name: page_number
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: page_size
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientListResponse'
  /v1/patientlist:
    get:
      tags:
        - Patient
      operationId: getPatientList
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientListResponse'
  /v1/patient/search:
    get:
      tags:
        - Patient
      operationId: searchPatients
      parameters:
        - name: searchString
          in: query
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "502":
          description: Bad Gateway
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PatientListResponse'
...
components:
  schemas:
    ErrorDto:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
        errorCode:
          type: integer
          format: int32
        errorText:
          type: string
...
    ContactDetailsDto:
      type: object
      properties:
        street:
          type: string
        postalCode:
          type: string
        country:
          type: string
        city:
          type: string
        primaryPhone:
          type: string
        secondaryPhone:
          type: string
    PatientDto:
      type: object
      properties:
        givenName:
          type: string
        familyName:
          type: string
        cpr:
          type: string
        customUserName:
          type: string
        patientContactDetails:
          $ref: '#/components/schemas/ContactDetailsDto'
        primaryRelativeName:
          type: string
        primaryRelativeAffiliation:
          type: string
        primaryRelativeContactDetails:
          $ref: '#/components/schemas/ContactDetailsDto'
        additionalRelativeContactDetails:
          type: array
          items:
            $ref: '#/components/schemas/ContactDetailsDto'
...
    CreatePatientRequest:
      type: object
      properties:
        patient:
          $ref: '#/components/schemas/PatientDto'
...
    PatientListResponse:
      type: object
      properties:
        patients:
          type: array
          items:
            $ref: '#/components/schemas/PatientDto'